fix: Money.multiply 이용시 음수가 담긴 Money가 나오는 문제 해결
- multiply 메소드는 validation을 하지 않기 때문에 음수가 담긴 Money 객체가 나오는 문제가 있었습니다.
- multiply 메소드는 유저들의 수익을 계산하기 위해 사용되고 있었고, 수익을 나타낼 때에는 음수를 포함하여야 하는 것이 더 적절하므로, Money 객체를 이용하지 않도록 결정하였습니다. 이에 관련 메소드들의 리턴 타입을 바꾸고, 필요없어진 multiply 객체는 삭제해였습니다.

+++ src/main/java/blackjack/domain/result/DealerResult.java
@@ -7,9 +7,9 @@ import java.util.Map;
 
 public class DealerResult {
     private final Map<ResultType, Integer> dealerResult;
-    private final long profit;
+    private final double profit;
 
-    DealerResult(Map<ResultType, Integer> dealerResult, long profit) {
+    DealerResult(Map<ResultType, Integer> dealerResult, double profit) {
         this.dealerResult = Collections.unmodifiableMap(dealerResult);
         this.profit = profit;
     }
@@ -18,7 +18,7 @@ public class DealerResult {
         return dealerResult.get(resultType);
     }
 
-    public long getProfit() {
+    public double getProfit() {
         return this.profit;
     }
 }
